Update Activity on Week 2
Asked about whether a designer should heavily invest on a certain user demographic, or try to keep it general, eg. for a software that garners to a superuser or a technical user, should the designer focus on making things easier for the superuser by putting all the functionality in the front, thus making it easier to access, but harder for general user to learn the software.